+++ ghost/core/package.json
@@ -118,7 +118,6 @@
     "@tryghost/members-api": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
-    "@tryghost/members-ssr": "0.0.0",
     "@tryghost/members-stripe-service": "0.0.0",
     "@tryghost/metrics": "1.0.37",
     "@tryghost/milestones": "0.0.0",
@@ -161,6 +160,7 @@
     "compression": "1.8.0",
     "connect-slashes": "1.4.0",
     "cookie-session": "2.1.0",
+    "cookies": "0.9.1",
     "cors": "2.8.5",
     "date-fns": "2.30.0",
     "dompurify": "3.2.3",
@@ -213,7 +213,7 @@
     "nconf": "0.12.1",
     "node-fetch": "2.7.0",
     "node-jose": "2.2.0",
-    "otplib" : "12.0.1",
+    "otplib": "12.0.1",
     "papaparse": "5.3.2",
     "path-match": "1.2.4",
     "probe-image-size": "7.2.3",
