--- ghost/members-ssr/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/members-ssr');
