--- ghost/members-ssr/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@tryghost/members-ssr",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "@types/cookies": "0.9.0",
-    "c8": "8.0.1",
-    "keypair": "1.0.4",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/debug": "0.1.32",
-    "@tryghost/errors": "1.3.5",
-    "cookies": "0.9.1",
-    "jsonwebtoken": "8.5.1"
-  }
-}
