--- ghost/members-ssr/test/hello.test.js
@@ -1,8 +1,0 @@
-require('should');
-
-describe('Hello world', function () {
-    it('Runs a test', function () {
-        // TODO: Write me!
-        'hello'.should.eql('hello');
-    });
-});
